change serie for series, and all related with this one.
