+++ MarvelApiClient/src/main/java/com/karumi/marvelapiclient/model/SeriesCollectionDto.java
@@ -1,0 +1,40 @@
+/*
+ *   Copyright (C) 2015 Karumi.
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+package com.karumi.marvelapiclient.model;
+
+import java.util.List;
+
+public class SeriesCollectionDto extends MarvelCollection<SeriesDto> {
+
+  public List<SeriesDto> getSeries() {
+    return getResults();
+  }
+
+  @Override public String toString() {
+    return "CharactersDto{"
+        + "offset="
+        + getOffset()
+        + ", limit="
+        + getLimit()
+        + ", total="
+        + getTotal()
+        + ", count="
+        + getCount()
+        + ", characters="
+        + getSeries().toString()
+        + '}';
+  }
+}
