--- MarvelApiClient/src/test/resources/getSerie.json
@@ -1,197 +1,0 @@
-{
-  "code": 200,
-  "status": "Ok",
-  "copyright": "© 2015 MARVEL",
-  "attributionText": "Data provided by Marvel. © 2015 MARVEL",
-  "attributionHTML": "<a href=\"http://marvel.com\">Data provided by Marvel. © 2015 MARVEL</a>",
-  "etag": "55122b303dfbc9fdfd12c080eded740b83354269",
-  "data": {
-    "offset": 0,
-    "limit": 20,
-    "total": 1,
-    "count": 1,
-    "results": [
-      {
-        "id": 15276,
-        "title": "5 Ronin (2011)",
-        "description": "description",
-        "resourceURI": "http://gateway.marvel.com/v1/public/series/15276",
-        "urls": [
-          {
-            "type": "detail",
-            "url": "http://marvel.com/comics/series/15276/5_ronin_2011?utm_campaign=apiRef&utm_source=838a08a2f4c39fa3fd218b1b2d43f19a"
-          }
-        ],
-        "startYear": 2011,
-        "endYear": 2011,
-        "rating": "+18",
-        "type": "",
-        "modified": "2011-05-20T16:25:29-0400",
-        "thumbnail": {
-          "path": "http://i.annihil.us/u/prod/marvel/i/mg/b/40/image_not_available",
-          "extension": "jpg"
-        },
-        "creators": {
-          "available": 8,
-          "collectionURI": "http://gateway.marvel.com/v1/public/series/15276/creators",
-          "items": [
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/creators/750",
-              "name": "David Aja",
-              "role": "penciller (cover)"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/creators/162",
-              "name": "John Cassaday",
-              "role": "penciller (cover)"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/creators/645",
-              "name": "Laurence Campbell",
-              "role": "penciller"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/creators/1421",
-              "name": "Tomm Coker",
-              "role": "penciller"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/creators/273",
-              "name": "Leandro Fernandez",
-              "role": "penciller"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/creators/10082",
-              "name": "Dalibor Talajic",
-              "role": "penciller"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/creators/94",
-              "name": "Peter Milligan",
-              "role": "writer"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/creators/4430",
-              "name": "Jeff Youngquist",
-              "role": "artist"
-            }
-          ],
-          "returned": 8
-        },
-        "characters": {
-          "available": 4,
-          "collectionURI": "http://gateway.marvel.com/v1/public/series/15276/characters",
-          "items": [
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/characters/1009268",
-              "name": "Deadpool"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/characters/1009351",
-              "name": "Hulk"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/characters/1009515",
-              "name": "Punisher"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/characters/1009718",
-              "name": "Wolverine"
-            }
-          ],
-          "returned": 4
-        },
-        "stories": {
-          "available": 12,
-          "collectionURI": "http://gateway.marvel.com/v1/public/series/15276/stories",
-          "items": [
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/stories/80038",
-              "name": "Interior #80038",
-              "type": "interiorStory"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/stories/80039",
-              "name": "5 Ronin #5",
-              "type": "cover"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/stories/80052",
-              "name": "5 RONIN 3 AJA COVER  ",
-              "type": "interiorStory"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/stories/80748",
-              "name": "5 Ronin #2",
-              "type": "cover"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/stories/80749",
-              "name": "5 Ronin #2",
-              "type": "interiorStory"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/stories/89879",
-              "name": "5 Ronin #1",
-              "type": "cover"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/stories/89880",
-              "name": "5 Ronin #1 ",
-              "type": "interiorStory"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/stories/90212",
-              "name": "5 Ronin #3 Aja Cover",
-              "type": "cover"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/stories/90349",
-              "name": "5 Ronin #4",
-              "type": "cover"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/stories/90404",
-              "name": "5 Ronin #5",
-              "type": "interiorStory"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/stories/93074",
-              "name": "5 Ronin HC",
-              "type": "cover"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/stories/93075",
-              "name": "5 Ronin HC DM Only Variant",
-              "type": "cover"
-            }
-          ],
-          "returned": 12
-        },
-        "comics": {
-          "available": 2,
-          "collectionURI": "http://gateway.marvel.com/v1/public/series/15276/comics",
-          "items": [
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/comics/41112",
-              "name": "5 Ronin (Hardcover)"
-            },
-            {
-              "resourceURI": "http://gateway.marvel.com/v1/public/comics/41113",
-              "name": "5 Ronin (Hardcover)"
-            }
-          ],
-          "returned": 2
-        },
-        "events": {
-          "available": 0,
-          "collectionURI": "http://gateway.marvel.com/v1/public/series/15276/events",
-          "items": [],
-          "returned": 0
-        },
-        "next": null,
-        "previous": null
-      }
-    ]
-  }
-}
