+++ sample/src/main/java/com/karumi/marvelapiclient/sample/MarveApiClientDemo.java
@@ -19,15 +19,15 @@ import com.karumi.marvelapiclient.CharacterApiClient;
 import com.karumi.marvelapiclient.ComicApiClient;
 import com.karumi.marvelapiclient.MarvelApiConfig;
 import com.karumi.marvelapiclient.MarvelApiException;
-import com.karumi.marvelapiclient.SerieApiClient;
+import com.karumi.marvelapiclient.SeriesApiClient;
 import com.karumi.marvelapiclient.model.CharacterDto;
 import com.karumi.marvelapiclient.model.CharactersDto;
 import com.karumi.marvelapiclient.model.CharactersQuery;
 import com.karumi.marvelapiclient.model.ComicDto;
 import com.karumi.marvelapiclient.model.ComicsDto;
 import com.karumi.marvelapiclient.model.MarvelResponse;
-import com.karumi.marvelapiclient.model.SerieDto;
 import com.karumi.marvelapiclient.model.SeriesDto;
+import com.karumi.marvelapiclient.model.SeriesCollectionDto;
 
 public class MarveApiClientDemo {
   public static void main(String[] args) {
@@ -61,12 +61,12 @@ public class MarveApiClientDemo {
         MarvelResponse<ComicDto> comic = comicApiClient.getComic("42882");
         System.out.println(comic.toString());
       } else if (operation.equals("getSeries")) {
-        SerieApiClient serieApiClient = new SerieApiClient(marvelApiConfig);
-        MarvelResponse<SeriesDto> series = serieApiClient.getAll(0, 10);
+        SeriesApiClient seriesApiClient = new SeriesApiClient(marvelApiConfig);
+        MarvelResponse<SeriesCollectionDto> series = seriesApiClient.getAll(0, 10);
         System.out.println(series.toString());
-      } else if (operation.equals("getSerie")) {
-        SerieApiClient serieApiClient = new SerieApiClient(marvelApiConfig);
-        MarvelResponse<SerieDto> serie = serieApiClient.getSerie("15276");
+      } else if (operation.equals("getSeriesById")) {
+        SeriesApiClient seriesApiClient = new SeriesApiClient(marvelApiConfig);
+        MarvelResponse<SeriesDto> serie = seriesApiClient.getSeriesById("15276");
         System.out.println(serie.toString());
       }
     } catch (MarvelApiException e) {
